+++ include/integrals.hpp
@@ -13,7 +13,7 @@
 /**
  * Given an operator type, an orbital basis and atoms, calculates the one-body integrals (associated to that operator type).
  */
-Eigen::MatrixXf compute_1body_integrals(const libint2::Operator&, const libint2::BasisSet&, const std::vector<libint2::Atom>&);
+Eigen::MatrixXd compute_1body_integrals(const libint2::Operator&, const libint2::BasisSet&, const std::vector<libint2::Atom>&);
 
 /**
  * Calculates the two-electron integrals, given an orbital basis and atoms.
